[WEBPORTFOLIO_012] feat: Install storybook complete
