+++ .storybook/main.cjs
@@ -1,3 +1,4 @@
+const { mergeConfig } = require('vite');
 module.exports = {
   "stories": [
     "../src/**/*.stories.mdx",
@@ -14,5 +15,13 @@ module.exports = {
   },
   "features": {
     "storyStoreV7": true
+  },
+  async viteFinal(config) {
+    return mergeConfig(config, {
+      resolve: (await import('../vite.config.js')).default.resolve,
+      optimizeDeps: {
+        include: ['storybook-dark-mode'],
+      }
+    });
   }
-}
+};
