+++ .storybook/preview.cjs
@@ -1,3 +1,5 @@
+import "./tailwind.css";
+import { mockServer } from "../src/mockServer/index";
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },
   controls: {
@@ -6,4 +8,4 @@ export const parameters = {
       date: /Date$/,
     },
   },
-}
+};
