+++ package.json
@@ -8,13 +8,15 @@
     "build": "vite build",
     "preview": "vite preview",
     "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
+    "build-storybook": "build-storybook",
+    "build-css": "npx tailwindcss -i ./src/utilities/index.css -o ./.storybook/tailwind.css --watch"
   },
   "dependencies": {
     "@fortawesome/fontawesome-svg-core": "^6.3.0",
     "@fortawesome/free-brands-svg-icons": "^6.3.0",
     "@fortawesome/free-solid-svg-icons": "^6.3.0",
     "@fortawesome/react-fontawesome": "^0.2.0",
+    "miragejs": "^0.1.47",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "vite-plugin-babel-macros": "^1.0.6"
