--- src/index.css
@@ -1,34 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Raleway:wght@400;600&display=swap');
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-@layer components {
-  .grid-layout {
-    @apply w-[90%] grid grid-cols-12 max-w-[1100px] h-full m-auto gap-[10px] overflow-hidden items-center
-  }
-
-  .hero-texts {
-    @apply pr-5 col-span-7
-  }
-
-  .btn-cv {
-    @apply w-[130px] inline-block text-center py-[10px] px-0 no-underline bg-eerie-black text-just-white
-  }
-
-  .parraph {
-    @apply mb-5 font-normal text-[1.4rem]
-  }
-
-  .about-us {
-    @apply h-auto pt-7 pb-14 bg-eerie-black text-just-white
-  }
-
-  .portfolio--icons {
-    @apply text-[25px] mt-0 ml-[5px] mb-0 mr-3
-  }
-
-  .button {
-    @apply w-[50%] max-w-[260px] h-[50px] min-w-[110px] my-[5px] mx-[10px] p-[5px] flex justify-center items-center text-[14px] rounded-xl bg-eerie-black no-underline text-just-white cursor-pointer hover:bg-just-white hover:text-eerie-black hover:border-eerie-black hover:border-2 focus:border-solid
-  }
-}
