+++ src/mockServer/index.js
@@ -1,0 +1,14 @@
+import { createServer } from "miragejs";
+
+export const mockServer = createServer({
+  routes() {
+    this.get("/buttons/:variant", (schema, request) => {
+      return {
+        data: {
+          text: `${request.params.variant}`,
+          variant: request.params.variant
+        }
+      };
+    });
+  }
+});
