+++ src/utilities/app.css
@@ -1,0 +1,42 @@
+* {
+  padding: 0;
+  margin: 0;
+  box-sizing: border-box;
+}
+
+[class^="fab fa"] {
+  font-size: 50px;
+  text-align: center;
+}
+
+/* Button from portfolio card */
+.button-container {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  border-bottom: 2px #151515 dotted;
+}
+
+.button-container [class^="fab fa"] {
+  font-size: 20px;
+  margin-left: 5px;
+  text-align: center;
+}
+
+/* Footer */
+[class^="footer-item"] a {
+  color: var(--just-white);
+  font-size: 50px;
+}
+
+[class^="footer-item"] a:hover {
+  color: var(--gray);
+}
+
+[class^="footer-item"] {
+  grid-column: span 3;
+}
+
+[class^="footer-item"]:hover {
+  border-bottom: 2px solid white;
+}
