+++ src/utilities/index.css
@@ -1,0 +1,28 @@
+@import url('https://fonts.googleapis.com/css2?family=Raleway:wght@400;600&display=swap');
+@import './buttonStyle.css';
+
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+@layer components {
+  .grid-layout {
+    @apply w-[90%] grid grid-cols-12 max-w-[1100px] h-full m-auto gap-[10px] overflow-hidden items-center
+  }
+
+  .hero-texts {
+    @apply pr-5 col-span-7
+  }
+
+  .parraph {
+    @apply mb-5 font-normal text-[1.4rem]
+  }
+
+  .about-us {
+    @apply h-auto pt-7 pb-14 bg-eerie-black text-just-white
+  }
+
+  .portfolio--icons {
+    @apply text-[25px] mt-0 ml-[5px] mb-0 mr-3
+  }
+}
